secure_doc_analyzer/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via kombu
asgiref==3.7.2
    # via django
async-timeout==5.0.1
    # via redis
bandit==1.8.2
    # via -r requirements.in
billiard==4.2.1
    # via celery
boto3==1.33.13
    # via -r requirements.in
botocore==1.33.13
    # via
    #   boto3
    #   s3transfer
celery==5.4.0
    # via -r requirements.in
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
coverage[toml]==7.6.11
    # via pytest-cov
django==3.2.25
    # via
    #   -r requirements.in
    #   django-auditlog
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-auditlog==3.0.0
    # via -r requirements.in
django-cors-headers==4.1.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.15.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.0
    # via -r requirements.in
exceptiongroup==1.2.2
    # via pytest
gunicorn==23.0.0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.4.2
    # via celery
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
packaging==24.2
    # via
    #   gunicorn
    #   pytest
pbr==6.1.1
    # via stevedore
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.50
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements.in
pygments==2.19.1
    # via rich
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pytest==8.3.4
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
pytest-cov==6.0.0
    # via -r requirements.in
pytest-django==4.9.0
    # via -r requirements.in
pytest-mock==3.14.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   django-auditlog
pytz==2025.1
    # via django
pyyaml==6.0.2
    # via bandit
redis==5.2.1
    # via -r requirements.in
rich==13.9.4
    # via bandit
s3transfer==0.8.2
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
stevedore==5.4.0
    # via bandit
tomli==2.2.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.7.1
    # via
    #   asgiref
    #   rich
tzdata==2025.1
    # via
    #   celery
    #   kombu
urllib3==1.26.20
    # via botocore
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
